Fixed VC warning
git-svn-id: http://svn.boost.org/svn/boost/trunk@18506 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/regex/v4/regex_token_iterator.hpp
@@ -91,7 +91,7 @@ public:
    {
       if(N == -1)
          return false;
-      if(N+1 < subs.size())
+      if(N+1 < (int)subs.size())
       {
          ++N;
          result =((subs[N] == -1) ? value_type(what.prefix().str()) : value_type(what[subs[N]].str()));
